feat: fix some colors and fonts
