+++ blocks/header/header.css
@@ -5,32 +5,59 @@
     display: flex;
     flex-direction: row;
     list-style-type: none;
-    gap: 1rem;
     align-items: center;
     border-bottom: 1px solid #ccc;
     background: white;
     z-index: 999;
-    transition: 0.5s ease-out all;
+    transition: 0.5s ease-out transform;
+    padding: 0 32px 0 31px;
 }
 
 .header button {
-    text-transform: none;
     box-shadow: none;
     border: 1px solid;
+    font-family: RMNeue, Arial, Helvetica, sans-serif;
+    font-size: 14px;
+    border-color: #3e31fa;
+    color: #3e31fa;
+    border-radius: 5px;
+    background-color: transparent;
+    flex: 0 0 auto;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    gap: 10px;
+    padding: 5px 10px;
+    text-transform: none;
 }
 
 .header.hidden {
     transform: translateY(calc(0rem - var(--nav-height)));
 }
 
+.main-menu {
+    display: flex;
+    flex: 0 0;
+    flex-direction: row;
+    list-style-type: none;
+    justify-content: space-around;
+    white-space: nowrap;
+    font-family: RMNeue, Arial, Helvetica, sans-serif;
+    font-size: 16px;
+    color: #43526e;
+    font-weight: bold;
+}
+
 .menu-item {
     border: 0;
     line-height: 1.2;
     color: #888;
+    padding: 10px 12px;
 }
 
 a.menu-item {
-    color: #888;
+    color: #43526E;
     text-decoration: none;
 }
 
@@ -115,3 +142,4 @@ header ul {
     margin: 0;
     padding: 0;
 }
+
