+++ blocks/header/header.js
@@ -1,94 +1,13 @@
-const DATA_URL = 'https://publish-p47754-e237356.adobeaemcloud.com/graphql/execute.json/realmadridmastersite/structurePage%3Balang=es-es';
-
-async function fetchData() {
-  try {
-    const response = await fetch(DATA_URL);
-    return await response.json();
-  } catch (error) {
-    console.error(error);
-  }
-  return null;
-}
-
-function createPopupMenu(data) {
-  const popupArea = document.createElement('div');
-  popupArea.classList.add('popup-menu-area');
-
-  const popup = document.createElement('div');
-  popup.classList.add('popup-menu');
-  popupArea.appendChild(popup);
-
-  popup.innerHTML = '<div class=".popup-background"></div>';
-
-  const hasChildrenIcon = `
-    <svg focusable="false" width="18" height="18" aria-hidden="true">
-      <use xlink:href="/blocks/header/cibeles-sprite.svg#chevron-right"></use>
-    </svg>`;
-
-  const subMenu = document.createElement('div');
-  subMenu.classList.add('sub-menu-area');
-
-  const imageArea = document.createElement('div');
-  imageArea.classList.add('image-menu-area');
-
-  const updateSubMenu = (index) => {
-    subMenu.innerHTML = data.data.header.items[0]
-      .mainNavigation[index].childNavigationItems.map((nav) => (
-        `<li><a class='menu-item' href="${nav.url}">${nav.title}</a></li>`
-      )).join('');
-    if (data.data.header.items[0].mainNavigation[index].image) {
-      // eslint-disable-next-line no-underscore-dangle
-      const imageUrl = data.data.header.items[0].mainNavigation[index].image._publishUrl;
-      imageArea.innerHTML = `
-        <img src="${imageUrl}"></img>
-      `;
-    } else {
-      imageArea.innerHTML = '';
-    }
-  };
-
-  updateSubMenu(data.data.header.items[0].mainNavigation
-    .findIndex((nav) => nav.childNavigationItems.length));
-
-  const mainMenu = document.createElement('ul');
-  mainMenu.classList.add('main-menu-area');
-  data.data.header.items[0].mainNavigation.forEach((nav, index) => {
-    const menuItem = document.createElement('li');
-    const link = document.createElement('a');
-    link.setAttribute('class', 'menu-item');
-    link.textContent = nav.title;
-    if (nav.childNavigationItems.length) {
-      link.innerHTML += hasChildrenIcon;
-      link.addEventListener('click', (e) => {
-        console.log(index);
-        e.preventDefault();
-        updateSubMenu(index);
-      });
-    } else {
-      link.setAttribute('href', nav.url);
-    }
-    menuItem.appendChild(link);
-    mainMenu.appendChild(menuItem);
-  });
-
-  popup.appendChild(mainMenu);
-  popup.appendChild(subMenu);
-  popup.appendChild(imageArea);
-
-  const footer = document.createElement('div');
-  popup.appendChild(footer);
-  footer.classList.add('footer-menu-area');
-  footer.innerHTML = '<button class="footer-menu-item">ES</button>';
-
-  return popupArea;
-}
+import fetchMenuData from './utils.js';
+import createPopupMenu from './popupMenu.js';
+import createMainMenu from './mainMenu.js';
 
 function addHamburger(block, data) {
   let state = false;
   const icon = () => {
     const src = `/blocks/header/cibeles-sprite.svg#${state ? 'times' : 'menu'}`;
     return `
-        <svg focusable="false" width="32" height="32" aria-hidden="true">
+        <svg focusable="false" width="24" height="24" aria-hidden="true">
             <use xlink:href="${src}"></use>
         </svg>
     `;
@@ -96,7 +15,7 @@ function addHamburger(block, data) {
   const popup = createPopupMenu(data);
   block.appendChild(popup);
   const menu = document.createElement('div');
-  menu.setAttribute('style', 'width: 32px; height: 32px; background-color: #fff; border-radius: 5px');
+  menu.setAttribute('style', 'background-color: #fff; border-radius: 5px; padding: 0px 9px; display: flex; flex-direction: column; justify-content: center; align-items: center');
   console.log(`menu button created, state: ${state}`);
   menu.innerHTML = icon();
   menu.addEventListener('click', () => {
@@ -111,15 +30,6 @@ function addHamburger(block, data) {
     }
   });
   block.appendChild(menu);
-}
-
-function createMenu(data) {
-  const menuItems = data.data.header.items[0].additionalNavigation.map((nav) => (
-    `<li><a class='menu-item' href="${nav.url}">${nav.title}</a></li>`
-  )).join('');
-  return `
-    <ul style='display: flex; flex: 0 0; flex-direction: row; list-style-type: none; gap: 1rem; justify-content: space-around; padding: 0; white-space: nowrap'>${menuItems}</ul>
-  `;
 }
 
 export default async function decorate(block) {
@@ -137,7 +47,7 @@ export default async function decorate(block) {
     lastScroll = window.scrollY;
   });
 
-  const data = await fetchData();
+  const data = await fetchMenuData();
   console.log(data.data.header.items[0]);
 
   addHamburger(block, data);
@@ -148,22 +58,22 @@ export default async function decorate(block) {
   const sponsorUrl = data.data.header.items[0].sponsors[1].logo._publishUrl;
 
   block.appendChild(document.createRange().createContextualFragment(`
-    <div style="flex: 1 0 auto; display: flex; flex-direction: row; justify-content: space-between; align-items: center; padding-right: 10px">
+    <div style="flex: 1 0 auto; display: flex; flex-direction: row; justify-content: space-between; align-items: center">
       <!-- Logos -->
       <div style="flex: 0 0 auto; display: flex; flex-direction: row; justify-content: space-between; align-items: center; gap: 10px">
         <svg focusable="false" width="40" height="40">
           <use xlink:href="/blocks/header/cibeles-sprite.svg#logo-rm"></use>
         </svg>
-        <div style="width: 1px; height: 40px; border-right: 1px solid #888"></div>
+        <div style="width: 1px; height: 32px; border-right: 1px solid #e1e5ea"></div>
         <img src='${logoUrl}' style="width: 40px; height: 40px; margin-left: 10px"/>
       </div>
-      ${createMenu(data)}
+      ${createMainMenu(data)}
       <div style="flex: 0; display: flex; flex-direction: row; justify-content: space-between; align-items: center; gap: 10px">
-        <img src='${sponsorUrl}' style="width: 40px; height: 40px; margin-left: 10px"/>
-        <svg focusable="false" width="24" height="24">
+        <img src='${sponsorUrl}' style="width: 57px; height: 40px; margin-left: 10px"/>
+        <svg focusable="false" width="24" height="24" style="filter: invert(75%) sepia(18%) saturate(182%) hue-rotate(178deg) brightness(95%) contrast(87%);">
           <use xlink:href="/blocks/header/cibeles-sprite.svg#dots-v"></use>
         </svg>
-        <button style="border-color: #3e31fa; color: #3e31fa; border-radius: 5px; background-color: transparent; flex: 0 0 auto; display: flex; flex-direction: row; justify-content: space-between; align-items: center; gap: 10px; padding: 5px 10px">
+        <button>
           <svg focusable="false" width="16" height="16" aria-hidden="true" style="filter: invert(26%) sepia(75%) saturate(7487%) hue-rotate(245deg) brightness(95%) contrast(107%);">
             <use xlink:href="/blocks/header/cibeles-sprite.svg#profile"></use>
           </svg>
